Small changes to benchmarking PR
