+++ src/ontime/module/__init__.py
@@ -1,2 +1,2 @@
-from . import preprocessing
+from . import processing
 from . import data
