+++ src/ontime/module/data/__init__.py
@@ -1,1 +1,1 @@
-from . import datasets
+from . import dataset
